+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 Now, a new command `cloudos cromwell` is available, with the following
 subcommands:
     * status
-    * restart
+    * start
     * stop
 
 ### 0.1.0 - 2022-07-07